Removed autobrightness correction in image... it tended to make the strips so bright that it would saturate the signal.
Furthermore, going back to determining the threshold by looking at the current control instead of by basing it off historic data obviates the need for image-wide brightness correction. It is very likely that this is the better way to go, because according to Kayla, the we need to use the control every time.