Changed strip colorbox recognition method to use green mask only, instead of red-green mask.
This is because the camera on the Moto 6 seemed to have higher red
content than the previous cameras used; and white portions of the strip
had a reddish hue. This shift to green mask only use was made with the
logic that the green mask is sufficient and necessary for detection, and
so adding red hues to this masks adds another area of potential failure.